fix(core): remove public set_read_counter_for_test and relocate exhaustion test to wacore
- Remove public set_read_counter_for_test mutator entirely from NoiseSocket so production code cannot mutate the read counter
- Relocate counter exhaustion test directly into wacore where it accesses the private read_counter Mutex directly under cfg(test)
